+++ samples/apps/smallbank/clients/small_bank_client.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#include "../smallbank_schema.h"
+#include "../smallbank_serializer.h"
 #include "perf_client.h"
 
 using namespace std;
