+++ samples/apps/smallbank/smallbank_serializer.h
